refactor: move ToggleSwitch to a new file
